trace: Fix parameter types in io
An upcoming patch will let the compiler warn us when we are silently
losing precision in traces.  In this case, the only platform where
pid_t in the caller does not match int in the trace definition is
64-bit mingw, where the system headers are still buggy in declaring
a 64-bit pid_t even though getpid() only returns 32 bits, so a cast
in the caller is the easiest workaround.

Message-Id: <20170313195547.21466-6-eblake@redhat.com>

+++ io/channel-command.c
@@ -39,7 +39,7 @@ qio_channel_command_new_pid(int writefd,
     ioc->writefd = writefd;
     ioc->pid = pid;
 
-    trace_qio_channel_command_new_pid(ioc, writefd, readfd, pid);
+    trace_qio_channel_command_new_pid(ioc, writefd, readfd, (int) pid);
     return ioc;
 }
